+++ edge-hub/src/Microsoft.Azure.Devices.Edge.Hub.Core/SubscriptionProcessor.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 namespace Microsoft.Azure.Devices.Edge.Hub.Core
 {
+    using System;
     using System.Collections.Concurrent;
     using System.Collections.Generic;
     using System.Linq;
@@ -8,43 +9,41 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core
     using System.Threading.Tasks.Dataflow;
     using Microsoft.Azure.Devices.Edge.Hub.Core.Cloud;
     using Microsoft.Azure.Devices.Edge.Hub.Core.Device;
+    using Microsoft.Azure.Devices.Edge.Hub.Core.Identity;
     using Microsoft.Azure.Devices.Edge.Util;
+    using Microsoft.Extensions.Logging;
+    using static System.FormattableString;
 
-    public interface ISubscriptionProcessor
-    {
-        Task AddSubscription(string id, DeviceSubscription deviceSubscription);
-
-        Task RemoveSubscription(string id, DeviceSubscription deviceSubscription);
-
-        Task ProcessSubscriptions(string id, IEnumerable<(DeviceSubscription, bool)> subscriptions);
-    }
-
-    public class SubscriptionProcessor
+    public class SubscriptionProcessor : ISubscriptionProcessor
     {
         readonly IConnectionManager connectionManager;
-        readonly ConcurrentDictionary<string, ConcurrentQueue<(DeviceSubscription, bool)>> subscriptionsToProcess;
+        readonly ConcurrentDictionary<string, ConcurrentQueue<(DeviceSubscription, bool)>> pendingSubscriptions;
         readonly ActionBlock<string> processSubscriptionsBlock;
         readonly IInvokeMethodHandler invokeMethodHandler;
 
-        public SubscriptionProcessor(IConnectionManager connectionManager, IInvokeMethodHandler invokeMethodHandler)
+        public SubscriptionProcessor(IConnectionManager connectionManager, IInvokeMethodHandler invokeMethodHandler, IDeviceConnectivityManager deviceConnectivityManager)
         {
+            Preconditions.CheckNotNull(deviceConnectivityManager, nameof(deviceConnectivityManager));
             this.connectionManager = Preconditions.CheckNotNull(connectionManager, nameof(connectionManager));
             this.invokeMethodHandler = Preconditions.CheckNotNull(invokeMethodHandler, nameof(invokeMethodHandler));
-            this.subscriptionsToProcess = new ConcurrentDictionary<string, ConcurrentQueue<(DeviceSubscription, bool)>>();
-            this.processSubscriptionsBlock = new ActionBlock<string>(this.ProcessSubscriptions);
+            this.pendingSubscriptions = new ConcurrentDictionary<string, ConcurrentQueue<(DeviceSubscription, bool)>>();
+            this.processSubscriptionsBlock = new ActionBlock<string>(this.ProcessPendingSubscriptions);
+            deviceConnectivityManager.DeviceConnected += this.DeviceConnected;
         }
 
         public Task AddSubscription(string id, DeviceSubscription deviceSubscription)
         {
+            Events.AddingSubscription(id, deviceSubscription);
             this.connectionManager.AddSubscription(id, deviceSubscription);
-            this.AddSubscriptionsToProcess(id, new List<(DeviceSubscription, bool)> { (deviceSubscription, true) });
+            this.AddToPendingSubscriptions(id, new List<(DeviceSubscription, bool)> { (deviceSubscription, true) });
             return Task.CompletedTask;
         }
 
         public Task RemoveSubscription(string id, DeviceSubscription deviceSubscription)
         {
+            Events.RemovingSubscription(id, deviceSubscription);
             this.connectionManager.RemoveSubscription(id, deviceSubscription);
-            this.AddSubscriptionsToProcess(id, new List<(DeviceSubscription, bool)> { (deviceSubscription, false) });
+            this.AddToPendingSubscriptions(id, new List<(DeviceSubscription, bool)> { (deviceSubscription, false) });
             return Task.CompletedTask;
         }
 
@@ -52,7 +51,7 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core
         {
             Preconditions.CheckNonWhiteSpace(id, nameof(id));
             List<(DeviceSubscription, bool)> subscriptionsList = Preconditions.CheckNotNull(subscriptions, nameof(subscriptions)).ToList();
-
+            Events.ProcessingSubscriptions(id, subscriptionsList);
             foreach ((DeviceSubscription deviceSubscription, bool addSubscription) in subscriptionsList)
             {
                 if (addSubscription)
@@ -65,48 +64,95 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core
                 }
             }
 
-            this.AddSubscriptionsToProcess(id, subscriptionsList);
+            this.AddToPendingSubscriptions(id, subscriptionsList);
             return Task.CompletedTask;
         }
 
         internal async Task ProcessSubscription(string id, Option<ICloudProxy> cloudProxy, DeviceSubscription deviceSubscription, bool addSubscription)
         {
-            switch (deviceSubscription)
+            Events.ProcessingSubscription(id, deviceSubscription);
+            try
             {
-                case DeviceSubscription.C2D:
-                    if (addSubscription)
-                    {
-                        cloudProxy.ForEach(c => c.StartListening());
-                    }
+                switch (deviceSubscription)
+                {
+                    case DeviceSubscription.C2D:
+                        if (addSubscription)
+                        {
+                            cloudProxy.ForEach(c => c.StartListening());
+                        }
 
-                    break;
+                        break;
 
-                case DeviceSubscription.DesiredPropertyUpdates:
-                    await cloudProxy.ForEachAsync(c => addSubscription ? c.SetupDesiredPropertyUpdatesAsync() : c.RemoveDesiredPropertyUpdatesAsync());
-                    break;
+                    case DeviceSubscription.DesiredPropertyUpdates:
+                        await cloudProxy.ForEachAsync(c => addSubscription ? c.SetupDesiredPropertyUpdatesAsync() : c.RemoveDesiredPropertyUpdatesAsync());
+                        break;
 
-                case DeviceSubscription.Methods:
-                    if (addSubscription)
-                    {
-                        await cloudProxy.ForEachAsync(c => c.SetupCallMethodAsync());
-                        await this.invokeMethodHandler.ProcessInvokeMethodSubscription(id);
-                    }
-                    else
-                    {
-                        await cloudProxy.ForEachAsync(c => c.RemoveCallMethodAsync());
-                    }
+                    case DeviceSubscription.Methods:
+                        if (addSubscription)
+                        {
+                            await cloudProxy.ForEachAsync(c => c.SetupCallMethodAsync());
+                            await this.invokeMethodHandler.ProcessInvokeMethodSubscription(id);
+                        }
+                        else
+                        {
+                            await cloudProxy.ForEachAsync(c => c.RemoveCallMethodAsync());
+                        }
 
-                    break;
+                        break;
 
-                case DeviceSubscription.ModuleMessages:
-                case DeviceSubscription.TwinResponse:
-                case DeviceSubscription.Unknown:
-                    // No Action required
-                    break;
+                    case DeviceSubscription.ModuleMessages:
+                    case DeviceSubscription.TwinResponse:
+                    case DeviceSubscription.Unknown:
+                        // No Action required
+                        break;
+                }
+            }
+            catch (Exception ex)
+            {
+                Events.ErrorProcessingSubscription(ex, id, deviceSubscription, addSubscription);
             }
         }
 
-        async Task ProcessSubscriptions(string id)
+        async void DeviceConnected(object sender, EventArgs eventArgs)
+        {
+            Events.DeviceConnectedProcessingSubscriptions();
+            try
+            {
+                IEnumerable<IIdentity> connectedClients = this.connectionManager.GetConnectedClients().ToList();
+                foreach (IIdentity identity in connectedClients)
+                {
+                    try
+                    {
+                        Events.ProcessingSubscriptions(identity);
+                        await this.ProcessExistingSubscriptions(identity.Id);
+                    }
+                    catch (Exception e)
+                    {
+                        Events.ErrorProcessingSubscriptions(e, identity);
+                    }
+                }
+            }
+            catch (Exception e)
+            {
+                Events.ErrorProcessingSubscriptions(e);
+            }
+        }
+
+        async Task ProcessExistingSubscriptions(string id)
+        {
+            Option<ICloudProxy> cloudProxy = await this.connectionManager.GetCloudConnection(id);
+            Option<IReadOnlyDictionary<DeviceSubscription, bool>> subscriptions = this.connectionManager.GetSubscriptions(id);
+            await subscriptions.ForEachAsync(
+                async s =>
+                {
+                    foreach (KeyValuePair<DeviceSubscription, bool> subscription in s)
+                    {
+                        await this.ProcessSubscription(id, cloudProxy, subscription.Key, subscription.Value);
+                    }
+                });
+        }
+
+        async Task ProcessPendingSubscriptions(string id)
         {
             ConcurrentQueue<(DeviceSubscription, bool)> clientSubscriptionsQueue = this.GetClientSubscriptionsQueue(id);
             if (!clientSubscriptionsQueue.IsEmpty)
@@ -119,7 +165,7 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core
             }
         }
 
-        void AddSubscriptionsToProcess(string id, List<(DeviceSubscription, bool)> subscriptions)
+        void AddToPendingSubscriptions(string id, List<(DeviceSubscription, bool)> subscriptions)
         {
             ConcurrentQueue<(DeviceSubscription, bool)> clientSubscriptionsQueue = this.GetClientSubscriptionsQueue(id);
             subscriptions.ForEach(s => clientSubscriptionsQueue.Enqueue(s));
@@ -127,6 +173,84 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core
         }
 
         ConcurrentQueue<(DeviceSubscription, bool)> GetClientSubscriptionsQueue(string id)
-            => this.subscriptionsToProcess.GetOrAdd(id, new ConcurrentQueue<(DeviceSubscription, bool)>());
+            => this.pendingSubscriptions.GetOrAdd(id, new ConcurrentQueue<(DeviceSubscription, bool)>());
+
+        static class Events
+        {
+            const int IdStart = HubCoreEventIds.SubscriptionProcessor;
+            static readonly ILogger Log = Logger.Factory.CreateLogger<SubscriptionProcessor>();
+
+            enum EventIds
+            {
+                ErrorProcessingSubscriptions = IdStart,
+                ErrorRemovingSubscription,
+                ErrorAddingSubscription,
+                AddingSubscription,
+                RemovingSubscription,
+                ProcessingSubscriptions,
+                ProcessingSubscription
+            }
+
+            public static void ErrorProcessingSubscriptions(Exception ex, IIdentity identity)
+            {
+                if (ex.HasTimeoutException())
+                {
+                    Log.LogDebug((int)EventIds.ErrorProcessingSubscriptions, ex, Invariant($"Timed out while processing subscriptions for client {identity.Id}. Will try again when connected."));
+                }
+                else
+                {
+                    Log.LogWarning((int)EventIds.ErrorProcessingSubscriptions, ex, Invariant($"Error processing subscriptions for client {identity.Id}."));
+                }
+            }
+
+            public static void ErrorProcessingSubscription(Exception ex, string id, DeviceSubscription subscription, bool addSubscription)
+            {
+                string operation = addSubscription ? "adding" : "removing";
+                if (ex.HasTimeoutException())
+                {
+                    Log.LogDebug((int)EventIds.ErrorAddingSubscription, ex, Invariant($"Timed out while {operation} subscription {subscription} for client {id}. Will try again when connected."));
+                }
+                else
+                {
+                    Log.LogWarning((int)EventIds.ErrorRemovingSubscription, ex, Invariant($"Error {operation} subscription {subscription} for client {id}."));
+                }
+            }
+
+            public static void AddingSubscription(string id, DeviceSubscription subscription)
+            {
+                Log.LogDebug((int)EventIds.AddingSubscription, Invariant($"Adding subscription {subscription} for client {id}."));
+            }
+
+            public static void RemovingSubscription(string id, DeviceSubscription subscription)
+            {
+                Log.LogDebug((int)EventIds.RemovingSubscription, Invariant($"Removing subscription {subscription} for client {id}."));
+            }
+
+            public static void ProcessingSubscriptions(IIdentity identity)
+            {
+                Log.LogInformation((int)EventIds.ProcessingSubscriptions, Invariant($"Processing subscriptions for client {identity.Id}."));
+            }
+
+            public static void ProcessingSubscription(string id, DeviceSubscription deviceSubscription)
+            {
+                Log.LogDebug((int)EventIds.ProcessingSubscription, Invariant($"Processing subscription {deviceSubscription} for client {id}."));
+            }
+
+            internal static void DeviceConnectedProcessingSubscriptions()
+            {
+                Log.LogInformation((int)EventIds.ProcessingSubscription, Invariant($"Device connected to cloud, processing subscriptions for connected clients."));
+            }
+
+            internal static void ErrorProcessingSubscriptions(Exception e)
+            {
+                Log.LogWarning((int)EventIds.ProcessingSubscription, e, Invariant($"Error processing subscriptions for connected clients."));
+            }
+
+            internal static void ProcessingSubscriptions(string id, List<(DeviceSubscription, bool)> subscriptionsList)
+            {
+                string subscriptions = String.Join(", ", subscriptionsList.Select(s => $"{s.Item1}"));
+                Log.LogInformation((int)EventIds.ProcessingSubscription, Invariant($"Processing subscriptions {subscriptions} for client {id}."));
+            }
+        }
     }
 }
