+++ edge-hub/src/Microsoft.Azure.Devices.Edge.Hub.Core/routing/RoutingEdgeHub.cs
@@ -39,7 +39,6 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Routing
             ITwinManager twinManager,
             string edgeDeviceId,
             IInvokeMethodHandler invokeMethodHandler,
-            IDeviceConnectivityManager deviceConnectivityManager,
             ISubscriptionProcessor subscriptionProcessor)
         {
             this.router = Preconditions.CheckNotNull(router, nameof(router));
@@ -48,7 +47,6 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Routing
             this.twinManager = Preconditions.CheckNotNull(twinManager, nameof(twinManager));
             this.edgeDeviceId = Preconditions.CheckNonWhiteSpace(edgeDeviceId, nameof(edgeDeviceId));
             this.invokeMethodHandler = Preconditions.CheckNotNull(invokeMethodHandler, nameof(invokeMethodHandler));
-            deviceConnectivityManager.DeviceConnected += this.DeviceConnected;
             this.subscriptionProcessor = Preconditions.CheckNotNull(subscriptionProcessor, nameof(subscriptionProcessor));
         }
 
@@ -126,22 +124,13 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Routing
         }
 
         public Task AddSubscription(string id, DeviceSubscription deviceSubscription)
-        {
-            Events.AddingSubscription(id, deviceSubscription);
-            return this.subscriptionProcessor.AddSubscription(id, deviceSubscription);
-        }
+            => this.subscriptionProcessor.AddSubscription(id, deviceSubscription);
 
         public Task RemoveSubscription(string id, DeviceSubscription deviceSubscription)
-        {
-            Events.RemovingSubscription(id, deviceSubscription);
-            return this.subscriptionProcessor.RemoveSubscription(id, deviceSubscription);
-        }
+            => this.subscriptionProcessor.RemoveSubscription(id, deviceSubscription);
 
         public Task ProcessSubscriptions(string id, IEnumerable<(DeviceSubscription, bool)> subscriptions)
-        {
-            Preconditions.CheckNonWhiteSpace(id, nameof(id));
-            return this.subscriptionProcessor.ProcessSubscriptions(id, Preconditions.CheckNotNull(subscriptions, nameof(subscriptions)));
-        }
+            => this.subscriptionProcessor.ProcessSubscriptions(id, subscriptions);
 
         public void Dispose()
         {
@@ -169,44 +158,6 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Routing
             }
         }
 
-        internal async Task ProcessSubscription(string id, Option<ICloudProxy> cloudProxy, DeviceSubscription deviceSubscription, bool addSubscription)
-        {
-            Events.ProcessingSubscription(id, deviceSubscription);
-            switch (deviceSubscription)
-            {
-                case DeviceSubscription.C2D:
-                    if (addSubscription)
-                    {
-                        cloudProxy.ForEach(c => c.StartListening());
-                    }
-
-                    break;
-
-                case DeviceSubscription.DesiredPropertyUpdates:
-                    await cloudProxy.ForEachAsync(c => addSubscription ? c.SetupDesiredPropertyUpdatesAsync() : c.RemoveDesiredPropertyUpdatesAsync());
-                    break;
-
-                case DeviceSubscription.Methods:
-                    if (addSubscription)
-                    {
-                        await cloudProxy.ForEachAsync(c => c.SetupCallMethodAsync());
-                        await this.invokeMethodHandler.ProcessInvokeMethodSubscription(id);
-                    }
-                    else
-                    {
-                        await cloudProxy.ForEachAsync(c => c.RemoveCallMethodAsync());
-                    }
-
-                    break;
-
-                case DeviceSubscription.ModuleMessages:
-                case DeviceSubscription.TwinResponse:
-                case DeviceSubscription.Unknown:
-                    // No Action required
-                    break;
-            }
-        }
-
         static void ValidateMessageSize(IRoutingMessage messageToBeValidated)
         {
             long messageSize = messageToBeValidated.Size();
@@ -230,45 +181,6 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Routing
             return routingMessage;
         }
 
-        async void DeviceConnected(object sender, EventArgs eventArgs)
-        {
-            Events.DeviceConnectedProcessingSubscriptions();
-            try
-            {
-                IEnumerable<IIdentity> connectedClients = this.connectionManager.GetConnectedClients().ToList();
-                foreach (IIdentity identity in connectedClients)
-                {
-                    try
-                    {
-                        Events.ProcessingSubscriptions(identity);
-                        await this.ProcessSubscriptions(identity.Id);
-                    }
-                    catch (Exception e)
-                    {
-                        Events.ErrorProcessingSubscriptions(e, identity);
-                    }
-                }
-            }
-            catch (Exception e)
-            {
-                Events.ErrorProcessingSubscriptions(e);
-            }
-        }
-
-        async Task ProcessSubscriptions(string id)
-        {
-            Option<ICloudProxy> cloudProxy = await this.connectionManager.GetCloudConnection(id);
-            Option<IReadOnlyDictionary<DeviceSubscription, bool>> subscriptions = this.connectionManager.GetSubscriptions(id);
-            await subscriptions.ForEachAsync(
-                async s =>
-                {
-                    foreach (KeyValuePair<DeviceSubscription, bool> subscription in s)
-                    {
-                        await this.ProcessSubscription(id, cloudProxy, subscription.Key, subscription.Value);
-                    }
-                });
-        }
-
         static class Events
         {
             const int IdStart = HubCoreEventIds.RoutingEdgeHub;
@@ -280,14 +192,7 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Routing
                 MessageReceived = 1501,
                 ReportedPropertiesUpdateReceived = 1502,
                 DesiredPropertiesUpdateReceived = 1503,
-                DeviceConnectionNotFound,
-                ErrorProcessingSubscriptions,
-                ErrorRemovingSubscription,
-                ErrorAddingSubscription,
-                AddingSubscription,
-                RemovingSubscription,
-                ProcessingSubscriptions,
-                ProcessingSubscription
+                DeviceConnectionNotFound
             }
 
             public static void MethodCallReceived(string fromId, string toId, string correlationId)
@@ -298,62 +203,6 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Routing
             public static void UnableToSendC2DMessageNoDeviceConnection(string id)
             {
                 Log.LogWarning((int)EventIds.DeviceConnectionNotFound, Invariant($"Unable to send C2D message to device {id} as an active device connection was not found."));
-            }
-
-            public static void ErrorProcessingSubscriptions(Exception ex, IIdentity identity)
-            {
-                if (ex.HasTimeoutException())
-                {
-                    Log.LogDebug((int)EventIds.ErrorProcessingSubscriptions, ex, Invariant($"Timed out while processing subscriptions for client {identity.Id}. Will try again when connected."));
-                }
-                else
-                {
-                    Log.LogWarning((int)EventIds.ErrorProcessingSubscriptions, ex, Invariant($"Error processing subscriptions for client {identity.Id}."));
-                }
-            }
-
-            public static void ErrorRemovingSubscription(Exception ex, string id, DeviceSubscription subscription)
-            {
-                if (ex.HasTimeoutException())
-                {
-                    Log.LogDebug((int)EventIds.ErrorAddingSubscription, ex, Invariant($"Timed out while removing subscription {subscription} for client {id}. Will try again when connected."));
-                }
-                else
-                {
-                    Log.LogWarning((int)EventIds.ErrorRemovingSubscription, ex, Invariant($"Error removing subscription {subscription} for client {id}."));
-                }
-            }
-
-            public static void ErrorAddingSubscription(Exception ex, string id, DeviceSubscription subscription)
-            {
-                if (ex.HasTimeoutException())
-                {
-                    Log.LogDebug((int)EventIds.ErrorAddingSubscription, ex, Invariant($"Timed out while adding subscription {subscription} for client {id}. Will try again when connected."));
-                }
-                else
-                {
-                    Log.LogDebug((int)EventIds.ErrorAddingSubscription, ex, Invariant($"Error adding subscription {subscription} for client {id}."));
-                }
-            }
-
-            public static void AddingSubscription(string id, DeviceSubscription subscription)
-            {
-                Log.LogDebug((int)EventIds.AddingSubscription, Invariant($"Adding subscription {subscription} for client {id}."));
-            }
-
-            public static void RemovingSubscription(string id, DeviceSubscription subscription)
-            {
-                Log.LogDebug((int)EventIds.RemovingSubscription, Invariant($"Removing subscription {subscription} for client {id}."));
-            }
-
-            public static void ProcessingSubscriptions(IIdentity identity)
-            {
-                Log.LogInformation((int)EventIds.ProcessingSubscriptions, Invariant($"Processing subscriptions for client {identity.Id}."));
-            }
-
-            public static void ProcessingSubscription(string id, DeviceSubscription deviceSubscription)
-            {
-                Log.LogInformation((int)EventIds.ProcessingSubscription, Invariant($"Processing subscription {deviceSubscription} for client {id}."));
             }
 
             public static void MessagesReceived(IIdentity identity, IList<IMessage> messages)
@@ -402,16 +251,6 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Routing
             {
                 Log.LogDebug((int)EventIds.DesiredPropertiesUpdateReceived, Invariant($"Desired properties update message received for {id ?? string.Empty}"));
             }
-
-            internal static void DeviceConnectedProcessingSubscriptions()
-            {
-                Log.LogInformation((int)EventIds.ProcessingSubscription, Invariant($"Device connected to cloud, processing subscriptions for connected clients."));
-            }
-
-            internal static void ErrorProcessingSubscriptions(Exception e)
-            {
-                Log.LogWarning((int)EventIds.ProcessingSubscription, e, Invariant($"Error processing subscriptions for connected clients."));
-            }
         }
 
         static class Metrics
@@ -435,7 +274,7 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Routing
 
             public static IDisposable MessageLatency(IIdentity identity) => Util.Metrics.Latency(GetTags(identity), EdgeHubMessageLatencyOptions);
 
-            internal static MetricTags GetTags(IIdentity identity)
+            static MetricTags GetTags(IIdentity identity)
             {
                 return new MetricTags("Id", identity.Id);
             }
